Migrate from android.icu.math.BigDecimal to java.math.BigDecimal to simplify formatting, conversions and future compatibility across all JVM during KMP migration. Use BigDecimal at all possible places instead of Double to prevent computation/rounding issues or precision losses.

+++ sdk/crypto-coindata/src/main/kotlin/cz/kotox/crypto/sdk/coindata/internal/data/database/entity/CoinMarketEntity.kt
@@ -2,6 +2,7 @@ package cz.kotox.crypto.sdk.coindata.internal.data.database.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
+import java.math.BigDecimal
 import kotlin.time.ExperimentalTime
 import kotlin.time.Instant
 
@@ -18,31 +19,31 @@ internal data class CoinMarketEntity(
     val symbol: String,
     val name: String,
     val imageUrl: String,
-    val currentPrice: Double,
-    val marketCap: Long,
+    val currentPrice: BigDecimal,
+    val marketCap: BigDecimal,
     val marketCapRank: Int,
-    val fullyDilutedValuation: Long?,
-    val totalVolume: Double,
-    val high24h: Double?,
-    val low24h: Double?,
-    val priceChange24h: Double?,
-    val priceChangePercentage24h: Double?,
-    val marketCapChange24h: Double?,
-    val marketCapChangePercentage24h: Double?,
-    val circulatingSupply: Double,
-    val totalSupply: Double?,
-    val maxSupply: Double?,
-    val ath: Double,
-    val athChangePercentage: Double,
+    val fullyDilutedValuation: BigDecimal?,
+    val totalVolume: BigDecimal,
+    val high24h: BigDecimal?,
+    val low24h: BigDecimal?,
+    val priceChange24h: BigDecimal?,
+    val priceChangePercentage24h: BigDecimal?,
+    val marketCapChange24h: BigDecimal?,
+    val marketCapChangePercentage24h: BigDecimal?,
+    val circulatingSupply: BigDecimal,
+    val totalSupply: BigDecimal?,
+    val maxSupply: BigDecimal?,
+    val ath: BigDecimal,
+    val athChangePercentage: BigDecimal,
     val athDate: Instant,
-    val atl: Double,
-    val atlChangePercentage: Double,
+    val atl: BigDecimal,
+    val atlChangePercentage: BigDecimal,
     val atlDate: Instant,
 
     // Flattened ROI (nullable)
-    val roiTimes: Double?,
+    val roiTimes: BigDecimal?,
     val roiCurrency: String?,
-    val roiPercentage: Double?,
+    val roiPercentage: BigDecimal?,
 
     val lastUpdated: Instant?,
     val cachedAt: Instant,
